tail_flas/requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
pdfplumber>=0.11
PyMuPDF>=1.24
pillow>=10.3
pyyaml>=6.0
plotly>=5.22
rapidfuzz>=3.9
python-dateutil>=2.9
opencv-python-headless>=4.10
pytesseract>=0.3.10